requirements.txt
transformers==4.35.0
urllib3==1.26.14

requests==2.28.2
pydantic==2.3.0
dynaconf==3.2.4
fastapi==0.103.1
uvicorn==0.23.2
pillow==9.2.0

cairosvg==2.7.1